feat: update basics package
