+++ go.mod
@@ -1,7 +1,7 @@
 module github.com/bnguyensn/go-hello
 
-go 1.16
+go 1.18
 
-require (
-	golang.org/x/tour v0.1.0
-)
+require golang.org/x/tour v0.1.0
+
+replace github.com/bnguyensn/go-hello/basics => ../basics
